+++ processor/src/main/java/com/github/gfx/android/orma/processor/exception/ProcessingException.java
@@ -13,7 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.gfx.android.orma.processor;
+package com.github.gfx.android.orma.processor.exception;
+
+import com.github.gfx.android.orma.processor.OrmaProcessor;
 
 import javax.lang.model.element.Element;
