+++ processor/src/main/java/com/github/gfx/android/orma/processor/generator/DatabaseWriter.java
@@ -13,9 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.gfx.android.orma.processor;
+package com.github.gfx.android.orma.processor.generator;
 
 import com.github.gfx.android.orma.annotation.OnConflict;
+import com.github.gfx.android.orma.processor.util.Annotations;
+import com.github.gfx.android.orma.processor.OrmaProcessor;
+import com.github.gfx.android.orma.processor.ProcessingContext;
+import com.github.gfx.android.orma.processor.model.SchemaDefinition;
+import com.github.gfx.android.orma.processor.util.Types;
 import com.squareup.javapoet.ClassName;
 import com.squareup.javapoet.CodeBlock;
 import com.squareup.javapoet.FieldSpec;
@@ -159,8 +164,8 @@ public class DatabaseWriter extends BaseWriter {
         Charset utf8 = Charset.forName("UTF-8");
 
         context.schemaMap.values().forEach(schema -> {
-            md.update(schema.createTableStatement.getBytes(utf8));
-            schema.createIndexStatements.forEach(statement -> md.update(statement.getBytes(utf8)));
+            md.update(schema.getCreateTableStatement().getBytes(utf8));
+            schema.getCreateIndexStatements().forEach(statement -> md.update(statement.getBytes(utf8)));
         });
 
         return bytesToHex(md.digest());
