+++ processor/src/main/java/com/github/gfx/android/orma/processor/generator/DeleterWriter.java
@@ -13,8 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.gfx.android.orma.processor;
+package com.github.gfx.android.orma.processor.generator;
 
+import com.github.gfx.android.orma.processor.ProcessingContext;
+import com.github.gfx.android.orma.processor.model.SchemaDefinition;
+import com.github.gfx.android.orma.processor.util.Types;
 import com.squareup.javapoet.MethodSpec;
 import com.squareup.javapoet.TypeSpec;
