+++ processor/src/main/java/com/github/gfx/android/orma/processor/generator/SelectorWriter.java
@@ -13,8 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.gfx.android.orma.processor;
 
+package com.github.gfx.android.orma.processor.generator;
+
+import com.github.gfx.android.orma.processor.model.ColumnDefinition;
+import com.github.gfx.android.orma.processor.ProcessingContext;
+import com.github.gfx.android.orma.processor.model.SchemaDefinition;
+import com.github.gfx.android.orma.processor.util.Strings;
+import com.github.gfx.android.orma.processor.util.Types;
 import com.squareup.javapoet.ClassName;
 import com.squareup.javapoet.MethodSpec;
 import com.squareup.javapoet.TypeSpec;
@@ -25,27 +31,27 @@ import java.util.stream.Stream;
 
 import javax.lang.model.element.Modifier;
 
-public class RelationWriter extends BaseWriter {
+public class SelectorWriter extends BaseWriter {
 
     private final SchemaDefinition schema;
 
     private final ConditionQueryHelpers conditionQueryHelpers;
 
-    public RelationWriter(ProcessingContext context, SchemaDefinition schema) {
+    public SelectorWriter(ProcessingContext context, SchemaDefinition schema) {
         super(context);
         this.schema = schema;
         this.conditionQueryHelpers = new ConditionQueryHelpers(context, schema, getTargetClassName());
     }
 
     ClassName getTargetClassName() {
-        return schema.getRelationClassName();
+        return schema.getSelectorClassName();
     }
 
     @Override
     public TypeSpec buildTypeSpec() {
         TypeSpec.Builder classBuilder = TypeSpec.classBuilder(getTargetClassName().simpleName());
         classBuilder.addModifiers(Modifier.PUBLIC);
-        classBuilder.superclass(Types.getRelation(schema.getModelClassName(), getTargetClassName()));
+        classBuilder.superclass(Types.getSelector(schema.getModelClassName(), getTargetClassName()));
 
         classBuilder.addMethods(buildMethodSpecs());
 
@@ -64,36 +70,15 @@ public class RelationWriter extends BaseWriter {
 
         methodSpecs.add(MethodSpec.constructorBuilder()
                 .addModifiers(Modifier.PUBLIC)
-                .addParameter(getTargetClassName(), "relation")
-                .addCode("super(relation);\n")
+                .addParameter(Types.getOrmaConditionBase(schema.getModelClassName()), "condition")
+                .addCode("super(condition);\n")
                 .build());
 
         methodSpecs.add(MethodSpec.methodBuilder("clone")
+                .addAnnotation(Override.class)
                 .addModifiers(Modifier.PUBLIC)
-                .addAnnotation(Annotations.override())
                 .returns(getTargetClassName())
                 .addStatement("return new $T(this)", getTargetClassName())
-                .build());
-
-        methodSpecs.add(MethodSpec.methodBuilder("selector")
-                .addAnnotations(Annotations.overrideAndNonNull())
-                .addModifiers(Modifier.PUBLIC)
-                .returns(schema.getSelectorClassName())
-                .addStatement("return new $T(this)", schema.getSelectorClassName())
-                .build());
-
-        methodSpecs.add(MethodSpec.methodBuilder("updater")
-                .addAnnotations(Annotations.overrideAndNonNull())
-                .addModifiers(Modifier.PUBLIC)
-                .returns(schema.getUpdaterClassName())
-                .addStatement("return new $T(this)", schema.getUpdaterClassName())
-                .build());
-
-        methodSpecs.add(MethodSpec.methodBuilder("deleter")
-                .addAnnotations(Annotations.overrideAndNonNull())
-                .addModifiers(Modifier.PUBLIC)
-                .returns(schema.getDeleterClassName())
-                .addStatement("return new $T(this)", schema.getDeleterClassName())
                 .build());
 
         methodSpecs.addAll(conditionQueryHelpers.buildConditionHelpers());
@@ -112,20 +97,18 @@ public class RelationWriter extends BaseWriter {
     }
 
     Stream<MethodSpec> buildOrderByHelpers(ColumnDefinition column) {
+        String columnName = column.getEscapedColumnName();
         return Stream.of(
                 MethodSpec.methodBuilder("orderBy" + Strings.toUpperFirst(column.name) + "Asc")
                         .addModifiers(Modifier.PUBLIC)
                         .returns(getTargetClassName())
-                        .addStatement("return orderBy($T.$L.orderInAscending())", schema.getSchemaClassName(),
-                                column.name)
+                        .addStatement("return orderBy($S)", columnName + " ASC")
                         .build(),
                 MethodSpec.methodBuilder("orderBy" + Strings.toUpperFirst(column.name) + "Desc")
                         .addModifiers(Modifier.PUBLIC)
                         .returns(getTargetClassName())
-                        .addStatement("return orderBy($T.$L.orderInDescending())", schema.getSchemaClassName(),
-                                column.name)
+                        .addStatement("return orderBy($S)", columnName + " DESC")
                         .build()
         );
     }
-
 }
