+++ processor/src/main/java/com/github/gfx/android/orma/processor/generator/UpdaterWriter.java
@@ -13,8 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.gfx.android.orma.processor;
+package com.github.gfx.android.orma.processor.generator;
 
+import com.github.gfx.android.orma.processor.util.Annotations;
+import com.github.gfx.android.orma.processor.model.AssociationDefinition;
+import com.github.gfx.android.orma.processor.model.ColumnDefinition;
+import com.github.gfx.android.orma.processor.ProcessingContext;
+import com.github.gfx.android.orma.processor.exception.ProcessingException;
+import com.github.gfx.android.orma.processor.model.SchemaDefinition;
+import com.github.gfx.android.orma.processor.util.Types;
 import com.squareup.javapoet.MethodSpec;
 import com.squareup.javapoet.ParameterSpec;
 import com.squareup.javapoet.TypeSpec;
@@ -105,7 +112,7 @@ public class UpdaterWriter extends BaseWriter {
                     );
                 }
 
-                SchemaDefinition modelSchema = context.getSchemaDef(r.modelType);
+                SchemaDefinition modelSchema = context.getSchemaDef(r.getModelType());
                 if (modelSchema == null) {
                     // FIXME: just stack errors and return in order to continue processing
                     throw new ProcessingException(Types.SingleAssociation.simpleName() + "<T> can handle only Orma models",
@@ -124,7 +131,7 @@ public class UpdaterWriter extends BaseWriter {
                                 .addModifiers(Modifier.PUBLIC)
                                 .returns(schema.getUpdaterClassName())
                                 .addParameter(
-                                        ParameterSpec.builder(r.modelType, column.name)
+                                        ParameterSpec.builder(r.getModelType(), column.name)
                                                 .addAnnotations(column.nullabilityAnnotations())
                                                 .build()
                                 )
