+++ processor/src/main/java/com/github/gfx/android/orma/processor/util/Strings.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.gfx.android.orma.processor;
+package com.github.gfx.android.orma.processor.util;
 
 public class Strings {
